remove submod "auth" and merge it to api cuz it was just a jwt parser when the jwt comes from a api request
